// frontend/src/gen/SudoResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SudoRetry } from "./SudoRetry";
import type { SudoState } from "./SudoState";

export type SudoResponse = { "kind": "ready", reauthentication: SudoState, } | { "kind": "confirmed", retry: SudoRetry | null, } | { "kind": "navigate", location: string, } | { "kind": "error", message: string, };